refactor: keep query renderer in Components
