+++ src/Apps/Notifications/NotificationsApp.tsx
@@ -1,5 +1,5 @@
-import { NotificationsQueryRenderer } from "./Components/Notifications"
 import { MetaTags } from "Components/MetaTags"
+import { NotificationsQueryRenderer } from "Components/Notifications/Notifications"
 
 export const NotificationsApp: React.FC = () => {
   return (
